by Patrick: add various casts and type corrections to make compilers feel more comfy
git-svn-id: svn://scm.orgis.org/mpg123/trunk@1893 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/libmpg123.c
@@ -71,7 +71,7 @@ static void frame_buffercheck(mpg123_handle *fr)
 	if(fr->firstoff && fr->num == fr->firstframe)
 	{
 		off_t byteoff = samples_to_bytes(fr, fr->firstoff);
-		if(fr->buffer.fill > byteoff)
+		if((off_t)fr->buffer.fill > byteoff)
 		{
 			fr->buffer.fill -= byteoff;
 			/* buffer.p != buffer.data only for own buffer */
@@ -89,7 +89,7 @@ static void frame_buffercheck(mpg123_handle *fr)
 	if(fr->lastoff && fr->num == fr->lastframe)
 	{
 		off_t byteoff = samples_to_bytes(fr, fr->lastoff);
-		if(fr->buffer.fill > byteoff)
+		if((off_t)fr->buffer.fill > byteoff)
 		{
 			fr->buffer.fill = byteoff;
 		}
